refactor: fetchNotices custom hook으로 분리 - #29
