+++ src/pages/ViewAllNotice/index.tsx
@@ -1,40 +1,20 @@
 import styled from 'styled-components';
-import { ChangeEvent, useCallback, useEffect, useState } from 'react';
+import { ChangeEvent, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from './Header';
 import NoticeCard from './NoticeCard';
-import { NoticeType, SortKey, SortOptions } from '../../types';
+import { SortKey, SortOptions } from '../../types';
 import { useAuthStore } from '../../store';
-import { getNotices } from '../../api/notice.ts';
+import useFetchNotices from '../../hooks/useFetchTotalNotice';
 
 const ViewAllNotice = () => {
-  const [notices, setNotices] = useState<NoticeType[]>([]);
   const [isSorted, setIsSorted] = useState<SortKey>('desc');
   const [page, setPage] = useState(0);
-  const [totalPage, setTotalPage] = useState(0);
-  const [isLoading, setIsLoading] = useState(false);
-  const sortOptions: SortOptions = { desc: '최신순', asc: '오래된순' };
-  const SIZE = 4;
   const navigate = useNavigate();
-
   const { role } = useAuthStore();
+  const sortOptions: SortOptions = { desc: '최신순', asc: '오래된순' };
 
-  const fetchNotices = useCallback(async () => {
-    try {
-      const response = await getNotices(isSorted, page, SIZE);
-      setTotalPage(response.totalPage);
-      setNotices(response.simpleAnnouncements);
-    } catch (err) {
-      alert('올바른 동작을 해주세요');
-      setIsLoading(true);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [isSorted, page]);
-
-  useEffect(() => {
-    fetchNotices();
-  }, [fetchNotices]);
+  const { notices, totalPage, isLoading } = useFetchNotices({ isSorted, page });
 
   const handleSort = (e: ChangeEvent<HTMLSelectElement>) => {
     setIsSorted(e.target.value as SortKey);
